API Respones changed
Using the Plugin it gets caught in an infinit loop waiting for the right response code. However the state is now changed from a numerical value to a string value.
I corrected it so, it can continue after recieving an "FINISHED" Value instead of waiting for a "2" which never comes.